chore(autoedit): Consolidate renderers
