+++ vscode/src/autoedits/autoedits-provider.ts
@@ -57,12 +57,10 @@ import type { AutoeditsUserPromptStrategy } from './prompt/base'
 import { createPromptProvider } from './prompt/create-prompt-provider'
 import { getCodeToReplaceData } from './prompt/prompt-utils/code-to-replace'
 import { getCurrentFilePath } from './prompt/prompt-utils/common'
-import { DefaultDecorator } from './renderer/decorators/default-decorator'
 import { InlineDiffDecorator } from './renderer/decorators/inline-diff-decorator'
 import { getAddedLines, getDecorationInfoFromPrediction } from './renderer/diff-utils'
 import { initImageSuggestionService } from './renderer/image-gen'
-import { AutoEditsInlineRendererManager } from './renderer/inline-manager'
-import { AutoEditsDefaultRendererManager, type AutoEditsRendererManager } from './renderer/manager'
+import { AutoEditsRendererManager } from './renderer/manager'
 import {
     extractAutoEditResponseFromCurrentDocumentCommentTemplate,
     shrinkReplacerTextToCodeToReplaceRange,
@@ -149,7 +147,6 @@ export type AutoeditClientCapabilities = Pick<
 >
 
 interface AutoeditsFeatures {
-    shouldRenderInline: boolean
     shouldHotStreak: boolean
     allowUsingWebSocket: boolean
 }
@@ -227,17 +224,11 @@ export class AutoeditsProvider implements vscode.InlineCompletionItemProvider, v
             allowUsingWebSocket: this.features.allowUsingWebSocket,
         })
 
-        this.rendererManager = this.features.shouldRenderInline
-            ? new AutoEditsInlineRendererManager(
-                  editor => new InlineDiffDecorator(editor),
-                  fixupController,
-                  this.requestManager
-              )
-            : new AutoEditsDefaultRendererManager(
-                  editor => new DefaultDecorator(editor),
-                  fixupController,
-                  this.requestManager
-              )
+        this.rendererManager = new AutoEditsRendererManager(
+            editor => new InlineDiffDecorator(editor),
+            fixupController,
+            this.requestManager
+        )
 
         this.onSelectionChangeDebounced = debounce(
             (event: vscode.TextEditorSelectionChangeEvent) => this.onSelectionChange(event),
@@ -681,12 +672,7 @@ export class AutoeditsProvider implements vscode.InlineCompletionItemProvider, v
             }
 
             if ('decorations' in renderOutput) {
-                await this.rendererManager.renderInlineDecorations(
-                    decorationInfo,
-                    renderOutput.decorations
-                )
-            } else if (renderOutput.type === 'legacy-decorations') {
-                await this.rendererManager.renderInlineDecorations(decorationInfo)
+                await this.rendererManager.renderInlineDecorations(renderOutput.decorations)
             }
 
             if ('inlineCompletionItems' in renderOutput) {
