+++ vscode/src/autoedits/utils.test.ts
@@ -147,54 +147,6 @@ describe('splitLinesKeepEnds', () => {
     })
 })
 
-describe('extractInlineCompletionFromRewrittenCode', () => {
-    it('handles basic case', () => {
-        const prediction = 'const prefix = 1;\nconst middle = 2;\nconst suffix = 3;'
-        const prefix = 'const prefix = 1;\n'
-        const suffix = '\nconst suffix = 3;'
-
-        const result = utils.extractInlineCompletionFromRewrittenCode(prediction, prefix, suffix)
-        expect(result).toBe('const middle = 2;')
-    })
-
-    it('handles empty prefix and suffix', () => {
-        const prediction = 'const x = 1;'
-        const prefix = ''
-        const suffix = ''
-
-        const result = utils.extractInlineCompletionFromRewrittenCode(prediction, prefix, suffix)
-        expect(result).toBe('const x = 1;')
-    })
-
-    it('handles multiline completion', () => {
-        const prediction = 'prefix\nline1\nline2\nline3\nsuffix'
-        const prefix = 'prefix\n'
-        const suffix = '\nsuffix'
-
-        const result = utils.extractInlineCompletionFromRewrittenCode(prediction, prefix, suffix)
-        expect(result).toBe('line1\nline2\nline3')
-    })
-
-    it('same line suffix test', () => {
-        const prediction = 'const prefix = 1;\nconst middle = 2;\nconst suffix = 3;'
-        const prefix = 'const prefix = 1;\n'
-        const suffix = 'middle = 2;\nconst suffix = 3;'
-
-        const result = utils.extractInlineCompletionFromRewrittenCode(prediction, prefix, suffix)
-        expect(result).toBe('const middle = 2;')
-    })
-
-    it('same line suffix test (with multiple lines)', () => {
-        const prediction =
-            'const prefix = 1;\nconst middle = 2;\nconst suffix = 3;\nconst superSuffix = 4;'
-        const prefix = 'const prefix = '
-        const suffix = '= 3;\nconst superSuffix = 4;'
-
-        const result = utils.extractInlineCompletionFromRewrittenCode(prediction, prefix, suffix)
-        expect(result).toBe('1;\nconst middle = 2;\nconst suffix = 3;')
-    })
-})
-
 describe('trimNewLineCharsFromString', () => {
     it('removes leading newlines', () => {
         const input = '\n\nHello World'
@@ -298,312 +250,6 @@ describe('isAllNewLineChars', () => {
         expect(utils.isAllNewLineChars('\t')).toBe(false)
         expect(utils.isAllNewLineChars(' \t\n')).toBe(false)
         expect(utils.isAllNewLineChars('\n\t ')).toBe(false)
-    })
-})
-
-describe('adjustPredictionIfInlineCompletionPossible', () => {
-    it('prediction when the prefix matches partially with suffix', () => {
-        const originalPrediction = '\n    private async func {\n'
-        const prefix = '\n    private '
-        const suffix = '\n\n    private async func {\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('returns original prediction if prefix or suffix not found', () => {
-        const originalPrediction = 'some code'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('adjusts the prediction when prefix and suffix are found and surrounding content is only new lines', () => {
-        const originalPrediction = '\n\nfunction test() {\n  console.log("Test");\n}\n\n'
-        const prefix = '\nfunction test() {'
-        const suffix = '}'
-        const expected = '\nfunction test() {\n  console.log("Test");\n}'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('returns original prediction if content before prefix is not all new lines', () => {
-        const originalPrediction = 'var a = 1;\nfunction test() {\n  console.log(a);\n}\n'
-        const prefix = 'function test() {'
-        const suffix = '}'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('returns original prediction if content after suffix is not all new lines', () => {
-        const originalPrediction = '\nfunction test() {\n  console.log("Test");\n}\nconsole.log("Done");'
-        const prefix = '\nfunction test() {'
-        const suffix = '}'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('handles cases where prefix and suffix are adjacent', () => {
-        const originalPrediction = '\n\nprefixsuffix\n\n'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const expected = 'prefixsuffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('trims new line characters correctly', () => {
-        const originalPrediction = '\n\n  content with new lines  \n\n'
-        const prefix = '\n  content'
-        const suffix = 'new lines  \n'
-        const expected = '\n  content with new lines  \n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('returns prediction when prefix is empty', () => {
-        const originalPrediction = '\nfunction test() {\n  console.log("Test");\n}\n'
-        const prefix = ''
-        const suffix = '}'
-        const expected = '\nfunction test() {\n  console.log("Test");\n}'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('returns original prediction when suffix is empty', () => {
-        const originalPrediction = '\nfunction test() {\n  console.log("Test");\n}\n'
-        const prefix = 'function test() {'
-        const suffix = ''
-        const expected = 'function test() {\n  console.log("Test");\n}\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles empty prefix and suffix', () => {
-        const originalPrediction = '\n\nSome content\n\n'
-        const prefix = ''
-        const suffix = ''
-        const expected = '\n\nSome content\n\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles strings with only newlines in original prediction, prefix, and suffix', () => {
-        const originalPrediction = '\n\n\n'
-        const prefix = '\n'
-        const suffix = '\n'
-        const expected = '\n\n\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles multiple occurrences of the suffix in the original prediction', () => {
-        const originalPrediction = '\nsome code\nsuffix\nmore code\nsuffix\n'
-        const prefix = 'some code\n'
-        const suffix = 'suffix\n'
-        const expected = 'some code\nsuffix\nmore code\nsuffix\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when original prediction is only newlines', () => {
-        const originalPrediction = '\n\n\n'
-        const prefix = '\n\n'
-        const suffix = '\n'
-        const expected = '\n\n\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when completion to extract is empty', () => {
-        const originalPrediction = 'prefixsuffix'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const expected = 'prefixsuffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when completion has only newlines and needs trimming', () => {
-        const originalPrediction = '\nprefix\n\n\nsuffix\n'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const expected = 'prefix\n\n\nsuffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when completion has only newlines and needs trimming and overlapping new lines in between', () => {
-        const originalPrediction = '\nprefix\n\n\nsuffix\n'
-        const prefix = 'prefix\n'
-        const suffix = '\n\nsuffix'
-        const expected = 'prefix\n\n\nsuffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when completion has only newlines and needs trimming and overlapping with extra new line chars', () => {
-        const originalPrediction = '\nprefix\n\n\nsuffix\n'
-        const prefix = 'prefix\n\n'
-        const suffix = '\n\nsuffix'
-        const expected = 'prefix\n\n\n\nsuffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when prefix and suffix are identical', () => {
-        const originalPrediction = '\nprefix\nmiddle content\nprefix\n'
-        const prefix = 'prefix'
-        const suffix = 'prefix'
-        const expected = 'prefix\nmiddle content\nprefix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when suffix has extra chars', () => {
-        const originalPrediction = '\nstart\nsuffix\nmiddle\nsuffix\nend\n'
-        const prefix = 'start\n'
-        const suffix = 'suffix\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('handles when suffix repeats multiple times in original prediction', () => {
-        const originalPrediction = '\nstart\nsuffix\nmiddle\nsuffix\n\n'
-        const prefix = 'start\n'
-        const suffix = 'suffix\n'
-        const expected = 'start\nsuffix\nmiddle\nsuffix\n'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('handles when all parameters are empty strings', () => {
-        const originalPrediction = ''
-        const prefix = ''
-        const suffix = ''
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe('')
-    })
-
-    it('handles when originalPrediction contains only newlines and spaces', () => {
-        const originalPrediction = '\n \n\n '
-        const prefix = '\n '
-        const suffix = '\n '
-        const expected = '\n \n\n '
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(expected)
-    })
-
-    it('returns original prediction when prediction before prefix contains non-newline characters', () => {
-        const originalPrediction = 'code before\nprefix\ncompletion\nsuffix'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
-    })
-
-    it('returns original prediction when prediction after suffix contains non-newline characters', () => {
-        const originalPrediction = 'prefix\ncompletion\nsuffix\ncode after'
-        const prefix = 'prefix'
-        const suffix = 'suffix'
-        const result = utils.adjustPredictionIfInlineCompletionPossible(
-            originalPrediction,
-            prefix,
-            suffix
-        )
-        expect(result).toBe(originalPrediction)
     })
 })
