+++ vscode/src/autoedits/utils.ts
@@ -28,19 +28,6 @@ export function splitLinesKeepEnds(s: string): string[] {
     return lines
 }
 
-export function extractInlineCompletionFromRewrittenCode(
-    prediction: string,
-    codeToRewritePrefix: string,
-    codeToRewriteSuffix: string
-): string {
-    const predictionWithoutPrefix = prediction.slice(codeToRewritePrefix.length)
-    const endIndex = predictionWithoutPrefix.length - codeToRewriteSuffix.length
-    const completion = predictionWithoutPrefix.slice(0, endIndex)
-    const completionNumLines = lines(completion).length
-    const completionWithSameLineSuffix = lines(predictionWithoutPrefix).slice(0, completionNumLines)
-    return completionWithSameLineSuffix.join(getNewLineChar(codeToRewritePrefix + codeToRewriteSuffix))
-}
-
 export function isDuplicatingTextFromRewriteArea({
     addedText,
     codeToReplaceData,
@@ -60,65 +47,6 @@ export function isDuplicatingTextFromRewriteArea({
     }
 
     return false
-}
-
-/**
- * Adjusts the prediction to enable inline completion when possible.
- *
- * This function attempts to modify the original prediction so that it can be seamlessly inserted
- * inline between the given prefix and suffix. The intent is to extract the relevant completion
- * content from the prediction by identifying where the prefix and suffix appear within it. It ensures
- * that any extra new line characters or unintended content outside the prefix and suffix are handled
- * appropriately to produce a clean inline completion.
- *
- * **High-level methodology:**
- * 1. **Strip New Line Characters:** Remove leading and trailing new line characters from the prefix and suffix to accurately match them within the prediction.
- * 2. **Find Indices:** Locate the indices of the prefix and suffix within the original prediction.
- *    - If either is not found, return the original prediction as adjustment is not possible.
- * 3. **Verify Surrounding Content:** Check that any content before the prefix and after the suffix in the prediction consists only of new line characters.
- *    - If there is other content, return the original prediction to avoid incorrect insertion.
- * 4. **Extract Completion Content:** Slice out the content between the prefix and suffix from the prediction.
- * 5. **Trim Excess New Lines:** Remove any unnecessary new line characters from the start and end of the completion content to align it properly with the surrounding
- */
-export function adjustPredictionIfInlineCompletionPossible(
-    originalPrediction: string,
-    prefix: string,
-    suffix: string
-): string {
-    const prefixWithoutNewLine = trimNewLineCharsFromString(prefix)
-    const suffixWithoutNewLine = trimNewLineCharsFromString(suffix)
-
-    const indexPrefix = originalPrediction.indexOf(prefixWithoutNewLine)
-    const indexSuffix = originalPrediction.lastIndexOf(suffixWithoutNewLine)
-    if (
-        indexPrefix === -1 ||
-        indexSuffix === -1 ||
-        indexPrefix + prefixWithoutNewLine.length > indexSuffix
-    ) {
-        return originalPrediction
-    }
-
-    const predictionBeforePrefixMatch = originalPrediction.slice(0, indexPrefix)
-    const predictionAfterSuffixMatch = originalPrediction.slice(
-        indexSuffix + suffixWithoutNewLine.length
-    )
-    if (
-        !isAllNewLineChars(predictionBeforePrefixMatch) ||
-        !isAllNewLineChars(predictionAfterSuffixMatch)
-    ) {
-        return originalPrediction
-    }
-
-    let completion = originalPrediction.slice(indexPrefix + prefixWithoutNewLine.length, indexSuffix)
-
-    const trimStartChars = Math.min(countNewLineCharsStart(completion), countNewLineCharsEnd(prefix))
-    const trimEndChars = Math.min(countNewLineCharsEnd(completion), countNewLineCharsStart(suffix))
-    completion = completion.substring(
-        trimStartChars,
-        Math.max(trimStartChars, completion.length - trimEndChars)
-    )
-    const prediction = prefix + completion + suffix
-    return prediction
 }
 
 /**
